Improve error messages
When there is a syntax error in a Tonel file, an error is raised. The problem is that when we load a project from disk, we do not know in which file the error is.

This change makes the TonelReader catch those exception to add the file info and pass the exception with more info. This will be helpful for debugging